# api/openapi.yaml
openapi: 3.1.0
info:
  title: agentctl API
  version: 0.1.0
  summary: Trace and control dangerous agent actions.
  description: |
    Schema-first contract for `agentctl`.

    The Go CLI is the current implementation, but this API document is the
    canonical cross-language contract for generated JS and Python clients.
servers:
  - url: http://localhost:8080
    description: Example local agentctl server
tags:
  - name: gate
  - name: trace
  - name: replay
paths:
  /v1/gate:
    post:
      tags: [gate]
      operationId: gateAction
      summary: Evaluate a risky action against policy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GateRequest"
      responses:
        "200":
          description: Decision returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"
  /v1/traces:
    get:
      tags: [trace]
      operationId: listTraces
      summary: List or search traces
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: session_id
          schema:
            type: string
        - in: query
          name: action
          schema:
            $ref: "#/components/schemas/Action"
        - in: query
          name: verdict
          schema:
            $ref: "#/components/schemas/Verdict"
        - in: query
          name: package
          schema:
            type: string
        - in: query
          name: since
          schema:
            type: string
            format: date-time
        - in: query
          name: until
          schema:
            type: string
            format: date-time
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Matching traces
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceListResponse"
  /v1/replay:
    post:
      tags: [replay]
      operationId: replaySession
      summary: Re-evaluate a recorded session with a given policy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplayRequest"
      responses:
        "200":
          description: Replay results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplayResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Action:
      type: string
      enum:
        - install_package
        - run_code
        - access_secret
        - write_file
        - call_external_api
    Verdict:
      type: string
      enum:
        - allow
        - deny
        - escalate
    RequestContext:
      type: object
      additionalProperties: false
      required:
        - session_id
        - timestamp
      properties:
        session_id:
          type: string
        model:
          type: string
        agent:
          type: string
        actor:
          type: string
        team:
          type: string
        turn:
          type: integer
        timestamp:
          type: string
          format: date-time
    InstallPackageParams:
      type: object
      additionalProperties: false
      required:
        - manager
        - package
      properties:
        manager:
          type: string
        package:
          type: string
        version:
          type: string
        hash:
          type: string
        pinned:
          type: boolean
    RunCodeParams:
      type: object
      additionalProperties: false
      required:
        - language
        - command
      properties:
        language:
          type: string
        command:
          type: string
        stdin:
          type: string
        network:
          type: boolean
    AccessSecretParams:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
        scope:
          type: string
        ttl:
          type: integer
    WriteFileParams:
      type: object
      additionalProperties: false
      required:
        - path
        - operation
      properties:
        path:
          type: string
        operation:
          type: string
        size_bytes:
          type: integer
          format: int64
    CallExternalAPIParams:
      type: object
      additionalProperties: false
      required:
        - url
        - method
      properties:
        url:
          type: string
          format: uri
        method:
          type: string
        domain:
          type: string
    ActionParams:
      oneOf:
        - $ref: "#/components/schemas/InstallPackageParams"
        - $ref: "#/components/schemas/RunCodeParams"
        - $ref: "#/components/schemas/AccessSecretParams"
        - $ref: "#/components/schemas/WriteFileParams"
        - $ref: "#/components/schemas/CallExternalAPIParams"
    GateRequest:
      type: object
      additionalProperties: false
      required:
        - action
        - params
        - reason
      properties:
        action:
          $ref: "#/components/schemas/Action"
        params:
          $ref: "#/components/schemas/ActionParams"
        reason:
          type: string
        context:
          $ref: "#/components/schemas/RequestContext"
    ActionRequest:
      allOf:
        - $ref: "#/components/schemas/GateRequest"
      description: Stored request shape inside a recorded decision.
    Decision:
      type: object
      additionalProperties: false
      required:
        - trace_id
        - verdict
        - risk_score
        - timestamp
        - request
        - reason
        - eval_duration_ms
      properties:
        trace_id:
          type: string
        verdict:
          $ref: "#/components/schemas/Verdict"
        risk_score:
          type: integer
          minimum: 0
          maximum: 100
        timestamp:
          type: string
          format: date-time
        request:
          $ref: "#/components/schemas/ActionRequest"
        reason:
          type: string
        matched_rules:
          type: array
          items:
            type: string
        approval_required:
          type: boolean
        approved_by:
          type: string
        eval_duration_ms:
          type: number
          format: double
    TraceListResponse:
      type: object
      additionalProperties: false
      required:
        - traces
      properties:
        traces:
          type: array
          items:
            $ref: "#/components/schemas/Decision"
    ReplayRequest:
      type: object
      additionalProperties: false
      required:
        - session_id
      properties:
        session_id:
          type: string
        policy_path:
          type: string
        limit:
          type: integer
          minimum: 1
    ReplayResponse:
      type: object
      additionalProperties: false
      required:
        - session_id
        - policy
        - results
      properties:
        session_id:
          type: string
        policy:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/Decision"